GPU mesh-part picking: 24+8 packed EntityIdBuffer (R32_UInt)
- Add MeshPartIdx to InstanceDescriptor (16 bytes, GPU-aligned)
- Consolidate InstanceDescriptor into common.fx (single source of truth)
- Pack entity ID (24-bit) + meshpart (8-bit) into R32_UInt EntityIdBuffer
- Zero render target bandwidth overhead vs original R32_Float
- Update all GBuffer/forward shaders to write packed uint
- Fix terrain descriptor buffer strides (12 -> 16 bytes)
- Initialize MeshPartIdx in terrain quadtree compute emitters

+++ Freefall.Engine/Resources/Shaders/cleargbuffer.fx
@@ -32,7 +32,7 @@ struct PSOutput
     float4 Normal : SV_Target1;
     float4 Data : SV_Target2;
     float4 Depth : SV_Target3;
-    float  EntityId : SV_Target4;
+    uint   EntityId : SV_Target4;
 };
 
 PSOutput PS(VSOutput input)
@@ -42,7 +42,7 @@ PSOutput PS(VSOutput input)
     output.Normal = float4(0,0,0,0);
     output.Data = float4(0,0,0,0);
     output.Depth = float4(0,0,0,0);
-    output.EntityId = 0.0;
+    output.EntityId = 0;
     return output;
 }
